tests/integration/manifests-tester/requirements.txt
ops
pyyaml
serialized-data-interface
